requirements.txt
numpy
moviepy
PyQt5
imutils
tensorflow
ffmpeg-python
pillow
matplotlib
opencv-python
scenedetect
dataclasses